Added custom hooks in the My Flashcard Page
