+++ src/app/App.tsx
@@ -11,7 +11,7 @@ import Dashboard from "../pages/Dashboard/Dashboard";
 import Profile from "../pages/Profile";
 import FlashcardMenu from "../pages/FlashcardMenu";
 import CreateFlashcards from "../pages/CreateFlashcards/CreateFlashcards";
-import MyFlashcards from "../pages/MyFlashcard/MyFlashcards";
+import MyFlashcards from "../pages/MyFlashcard/FlashcardSetList/MyFlashcards";
 import FlashcardSet from "../pages/FlashcardSet/FlashcardSet";
 import PopularFlashcards from "../pages/PopularFlashcards/PopularFlashcards";
 import PopularFlashcardSet from "../pages/PopularFlashcards/PopularFlashcardSet";
